10.07.20.02

+++ school/csci/assn04/bitsum.cpp
@@ -2,13 +2,13 @@
 #include "bitsum.h"
 
 int numberOfOnes(unsigned long int value) {
-
   int rem = 0;
   int numOfOnes = 0;
-  
+
   while (value >= 1) {
     rem = value % 2;
-    value = (int) value / 2;
+    value =  value / 2;
+   // std::cout << "value: "<< value << std::endl;
     if (rem == 1) {
       numOfOnes++;
     } //if
